AppContextPluginAliasSetLikeDialogComponentContextXXX was PluginAliasSetLikeDialogComponentContextBasicXXX
